Fix Clang-15 warnings (#831)
Redundant casts, noexcept for move constructors, and usage of default and delete make the code more performant and maintainable.

+++ include/boost/math/special_functions/ulp.hpp
@@ -26,7 +26,7 @@ T ulp_imp(const T& val, const std::true_type&, const Policy& pol)
 
    int fpclass = (boost::math::fpclassify)(val);
 
-   if(fpclass == (int)FP_NAN)
+   if(fpclass == FP_NAN)
    {
       return policies::raise_domain_error<T>(
          function,
@@ -60,13 +60,13 @@ T ulp_imp(const T& val, const std::false_type&, const Policy& pol)
 
    int fpclass = (boost::math::fpclassify)(val);
 
-   if(fpclass == (int)FP_NAN)
+   if(fpclass == FP_NAN)
    {
       return policies::raise_domain_error<T>(
          function,
          "Argument must be finite, but got %1%", val, pol);
    }
-   else if((fpclass == (int)FP_INFINITE) || (fabs(val) >= tools::max_value<T>()))
+   else if((fpclass == FP_INFINITE) || (fabs(val) >= tools::max_value<T>()))
    {
       return (val < 0 ? -1 : 1) * policies::raise_overflow_error<T>(function, nullptr, pol);
    }
